Implementing Enum

+++ MainUI.cs
@@ -236,11 +236,12 @@ namespace BudgetSaverApp
         /// <summary>
         /// Loads data in the Stats tab.
         /// </summary>
-        /// <param name="stats"></param>
+        /// <param name="sta
+        /// 
         private void SetStatsInfo(Stats stats, string setting)
         {
             LabelStatsWeeklyTransactionAmount.Text = setting + "transaction amount: " + stats.TransactionAmount;
-            LabelStatsWeeklyIncome.Text = setting + "income: " + stats.Income + " €";
+            LabelStatsWeeklyIncome.Text = setting + "income: " + ((double)stats.Income) + " €";
             LabelStatsWeeklyExpenses.Text = setting + "expenses: " + stats.Expenses + " €";
             LabelStatsFrequentCategory.Text = setting + "most frequent category: " + stats.FrequentCategory;
             LabelStatsWeeklyBalance.Text = setting + "balance: " + (stats.Income - stats.Expenses) + " €";
